# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Twilight Core Explorer API",
    "description": "Read-only public API for the Twilight Core explorer (Phase 9a foundation).",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/txs": {
      "get": {
        "summary": "List transactions (newest first)",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "pattern": "^\\d+$",
              "type": "string"
            },
            "in": "query",
            "name": "height",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "page"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "hash",
                          "height",
                          "index",
                          "status",
                          "code",
                          "gasUsed",
                          "gasWanted",
                          "memo",
                          "messageTypes",
                          "signerAddresses"
                        ],
                        "properties": {
                          "hash": {
                            "type": "string"
                          },
                          "height": {
                            "pattern": "^\\d+$",
                            "description": "Decimal height/id as string",
                            "type": "string"
                          },
                          "index": {
                            "type": "integer"
                          },
                          "status": {
                            "type": "string"
                          },
                          "code": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "gasUsed": {
                            "anyOf": [
                              {
                                "pattern": "^\\d+$",
                                "description": "Decimal height/id as string",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "gasWanted": {
                            "anyOf": [
                              {
                                "pattern": "^\\d+$",
                                "description": "Decimal height/id as string",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "memo": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "messageTypes": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "signerAddresses": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        },
                        "title": "TxListItem"
                      }
                    },
                    "page": {
                      "type": "object",
                      "required": [
                        "limit",
                        "nextCursor"
                      ],
                      "properties": {
                        "limit": {
                          "minimum": 1,
                          "type": "integer"
                        },
                        "nextCursor": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/txs/{hash}": {
      "get": {
        "summary": "Get a transaction by hash",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "raw"
              ]
            },
            "in": "query",
            "name": "include",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "hash",
                        "height",
                        "index",
                        "status",
                        "code",
                        "gasUsed",
                        "gasWanted",
                        "memo",
                        "messageTypes",
                        "signerAddresses",
                        "time",
                        "fee",
                        "messages",
                        "events"
                      ],
                      "properties": {
                        "hash": {
                          "type": "string"
                        },
                        "height": {
                          "pattern": "^\\d+$",
                          "description": "Decimal height/id as string",
                          "type": "string"
                        },
                        "index": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "code": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "gasUsed": {
                          "anyOf": [
                            {
                              "pattern": "^\\d+$",
                              "description": "Decimal height/id as string",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "gasWanted": {
                          "anyOf": [
                            {
                              "pattern": "^\\d+$",
                              "description": "Decimal height/id as string",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "memo": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "messageTypes": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "signerAddresses": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "time": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "fee": {
                          "anyOf": [
                            {},
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "messages": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "msgIndex",
                              "typeUrl",
                              "module",
                              "typeName",
                              "decodedJson",
                              "decodeError"
                            ],
                            "properties": {
                              "msgIndex": {
                                "type": "integer"
                              },
                              "typeUrl": {
                                "type": "string"
                              },
                              "module": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "typeName": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "decodedJson": {
                                "anyOf": [
                                  {},
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "decodeError": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "raw": {}
                            },
                            "title": "Message"
                          }
                        },
                        "events": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "phase",
                              "type",
                              "msgIndex",
                              "eventIndex",
                              "attributes"
                            ],
                            "properties": {
                              "phase": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              },
                              "msgIndex": {
                                "anyOf": [
                                  {
                                    "type": "integer"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "eventIndex": {
                                "type": "integer"
                              },
                              "attributes": {}
                            },
                            "title": "Event"
                          }
                        },
                        "raw": {
                          "type": "object",
                          "required": [
                            "tx",
                            "result"
                          ],
                          "properties": {
                            "tx": {},
                            "result": {}
                          }
                        }
                      },
                      "title": "TxDetail"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}